adds hour, min, & sec hands
