+++ client/src/components/Clock/Clock.jsx
@@ -1,6 +1,6 @@
 import "./Clock.css";
 import Numbers from './Numbers/Numbers';
-import GetTime from "./ClockHands/GetTime";
+import ClockHands from "./ClockHands/ClockHands";
 import SubmitForm from "../SubmitForm/SubmitForm";
 
 const Clock = ({ setSubmittedZip }) => {
@@ -8,7 +8,7 @@ const Clock = ({ setSubmittedZip }) => {
 	return (
 		<div className="clock-frame">
             <Numbers />
-            <GetTime />
+            <ClockHands />
 			<SubmitForm setSubmittedZip={(zipCode) => setSubmittedZip(zipCode)} />
         </div>
     )
