+++ client/src/components/Clock/ClockHands/ClockHands.css
@@ -1,27 +1,35 @@
 .clock-hands {
-    position: absolute;
+    position: relative;
     z-index: 0;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    bottom: 97%;
     width: 100%;
     height: 100%;
     border-radius: 50%;
     /* border: 5px solid rgb(179, 51, 51); */
 }
 
+#center-point {
+    /* visibility: hidden; */
+    position: relative;
+    width: 100%;
+    height: 100%;
+    z-index: 1;
+    /* border: 3px solid green; */
+}
+
 #inner-center-point {
+    z-index: 1;
     width: 4%;
     height: 4%;
     border-radius: 50%;
     position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
+    left: 48%;
+    top: 48%;
     background-color: rgb(89, 89, 89);
 }
 
 #outer-center-point {
+    z-index: 1;
     width: 7%;
     height: 7%;
     border-radius: 50%;
@@ -32,18 +40,34 @@
     background-color: rgb(170, 170, 170);
 }
 
-.hour-hand {
-    position: relative;
-    background-color: black;
+#hour-hand {
+    position: absolute;
+    background-color: rgb(255, 0, 0);
     width: 1%;
-    height: 40%;
-    /* bottom: 50%; */
+    height: 25%;
+    left: 50%;
+    bottom: 50%;
+    transform-origin: bottom left;
+    transform: translate(-48%, 0%);
+    /* transform: rotate(235deg); */
 }
 
-/* #test {
-    color: red;
+#minute-hand {
     position: absolute;
-    bottom: 50%;
+    background-color: rgb(13, 0, 255);
+    width: 1%;
+    height: 32%;
     left: 50%;
-    transform: translate(-50%, -50%);
-} */
+    bottom: 50%;
+    transform: translate(-48%, 0%);
+}
+
+#second-hand {
+    position: absolute;
+    background-color: rgb(255, 115, 0);
+    width: 0.5%;
+    height: 35%;
+    left: 50%;
+    bottom: 50%;
+    transform: translate(-48%, 0%);
+}
